Moving Functions from rust-gcc-diagnostics to rust-diagnostics.cc
gcc/rust/ChangeLog:

	* Make-lang.in: Modified
	* rust-diagnostics.cc (rust_be_get_quotechars): Added
		(rust_be_internal_error_at): Added
		(rust_be_error_at): Added
		(class rust_error_code_rule): Added
		(rust_be_warning_at): Added
		(rust_be_fatal_error): Added
		(rust_be_inform): Added
		(rust_be_debug_p): Added

	* rust-gcc-diagnostics.cc: Removed
