+++ gcc/rust/Make-lang.in
@@ -66,7 +66,6 @@ GRS_OBJS = \
     rust/rust-lang.o \
     rust/rust-object-export.o \
     rust/rust-linemap.o \
-    rust/rust-gcc-diagnostics.o \
     rust/rust-diagnostics.o \
     rust/rust-gcc.o \
     rust/rust-token.o \
